# services/api/internal/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: 初芽记 chuyaji API
  version: "0.2.0"
  description: |
    宝宝 / 宝妈双工作台 HTTP API（Gin + SQLite）。

    认证方式：除登录、健康检查、公开附件外，请求头需携带 `Authorization: Bearer <JWT>`。
    开发模式：当服务端 `CHUYAJI_DEV_MODE=true` 时，登录 body 可使用 `{"code":"dev"}`。
servers:
  - url: http://127.0.0.1:8282
    description: 本地默认
  - url: /
    description: 相对当前主机与端口
...
paths:
...
  /api/v1/records/{id}/attachments/upload:
    post:
      tags: [Attachment]
      summary: 上传宝宝记录附件
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: 已上传
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
  /api/v1/mother-records/{id}/attachments/upload:
    post:
      tags: [Attachment]
      summary: 上传宝妈记录附件
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: 已上传
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    idPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    Attachment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        owner_type:
          type: string
          enum: [baby_record, mother_record]
        owner_id:
          type: integer
          format: int64
        url:
          type: string
        thumb_url:
          type: string
        sort_order:
          type: integer
        size:
          type: integer
          format: int64
...